# api/templates/openapi.json
{
   "openapi": "3.1.0",
   "info": {
      "title": "biodata-registry-service",
      "description": "\n## biodata-registry-service\n\nService to fetch data from Registry.\n\n",
      "version": "0.0.0"
   },
   "paths": {
...
      "/quality_control": {
         "post": {
            "tags": [
               "core"
            ],
            "summary": "Create Quality Control",
            "operationId": "create_quality_control",
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "$ref": "#/components/schemas/QualityControlCreate"
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/QualityControls"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "get": {
            "tags": [
               "core"
            ],
            "summary": "Get Quality Control",
            "operationId": "get_quality_control",
            "parameters": [
               {
                  "name": "id",
                  "in": "query",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "title": "Id"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/QualityControls"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "delete": {
            "tags": [
               "core"
            ],
            "summary": "Delete",
            "operationId": "delete_quality_control",
            "parameters": [
               {
                  "name": "id",
                  "in": "query",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "title": "Id"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {}
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "put": {
            "tags": [
               "core"
            ],
            "summary": "Update",
            "operationId": "update_quality_control",
            "parameters": [
               {
                  "name": "id",
                  "in": "query",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "title": "Id"
                  }
               }
            ],
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "$ref": "#/components/schemas/QualityControlUpdate"
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/QualityControls"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/quality_controls": {
         "get": {
            "tags": [
               "core"
            ],
            "summary": "Get Quality Controls",
            "operationId": "get_quality_controls",
            "parameters": [
               {
                  "name": "offset",
                  "in": "query",
                  "required": false,
                  "schema": {
                     "type": "integer",
                     "default": 0,
                     "title": "Offset"
                  }
               },
               {
                  "name": "limit",
                  "in": "query",
                  "required": false,
                  "schema": {
                     "type": "integer",
                     "maximum": 1000,
                     "default": 10,
                     "title": "Limit"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "array",
                           "items": {
                              "$ref": "#/components/schemas/QualityControls"
                           },
                           "title": "Response Get Quality Controls"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
...
      "/schema": {
         "post": {
            "tags": [
               "core"
            ],
            "summary": "Create Schema",
            "operationId": "create_schema",
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "$ref": "#/components/schemas/SchemaCreate"
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Schemas"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "get": {
            "tags": [
               "core"
            ],
            "summary": "Get Schema",
            "operationId": "get_schema",
            "parameters": [
               {
                  "name": "id",
                  "in": "query",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "title": "Id"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Schemas"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "delete": {
            "tags": [
               "core"
            ],
            "summary": "Delete",
            "operationId": "delete_schema",
            "parameters": [
               {
                  "name": "id",
                  "in": "query",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "title": "Id"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {}
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "put": {
            "tags": [
               "core"
            ],
            "summary": "Update",
            "operationId": "update_schema",
            "parameters": [
               {
                  "name": "id",
                  "in": "query",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "title": "Id"
                  }
               }
            ],
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "$ref": "#/components/schemas/SchemaUpdate"
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Schemas"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/schemas": {
         "get": {
            "tags": [
               "core"
            ],
            "summary": "Get Schemas",
            "operationId": "get_schemas",
            "parameters": [
               {
                  "name": "offset",
                  "in": "query",
                  "required": false,
                  "schema": {
                     "type": "integer",
                     "default": 0,
                     "title": "Offset"
                  }
               },
               {
                  "name": "limit",
                  "in": "query",
                  "required": false,
                  "schema": {
                     "type": "integer",
                     "maximum": 1000,
                     "default": 10,
                     "title": "Limit"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "array",
                           "items": {
                              "$ref": "#/components/schemas/Schemas"
                           },
                           "title": "Response Get Schemas"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
...
   },
   "components": {
      "schemas": {
...
         "HTTPValidationError": {
            "properties": {
               "detail": {
                  "items": {
                     "$ref": "#/components/schemas/ValidationError"
                  },
                  "type": "array",
                  "title": "Detail"
               }
            },
            "type": "object",
            "title": "HTTPValidationError"
         },
...
         "QualityControlCreate": {
            "properties": {
               "data": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Data"
               },
               "schema_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Schema Id"
               },
               "space_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Space Id"
               },
               "data_asset_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Data Asset Id"
               }
            },
            "type": "object",
            "title": "QualityControlCreate"
         },
         "QualityControlUpdate": {
            "properties": {
               "data": {
                  "anyOf": [
                     {
                        "additionalProperties": true,
                        "type": "object"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Data"
               },
               "schema_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Schema Id"
               },
               "space_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Space Id"
               },
               "data_asset_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Data Asset Id"
               }
            },
            "type": "object",
            "title": "QualityControlUpdate"
         },
         "QualityControls": {
            "properties": {
               "id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Id"
               },
               "data": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Data"
               },
               "schema_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Schema Id"
               },
               "space_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Space Id"
               },
               "data_asset_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Data Asset Id"
               }
            },
            "type": "object",
            "title": "QualityControls"
         },
         "SchemaCreate": {
            "properties": {
               "name": {
                  "type": "string",
                  "maxLength": 50,
                  "title": "Name"
               },
               "version": {
                  "type": "string",
                  "maxLength": 50,
                  "title": "Version"
               },
               "data": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Data"
               },
               "schema_entity_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Schema Entity Id"
               }
            },
            "type": "object",
            "required": [
               "name",
               "version"
            ],
            "title": "SchemaCreate"
         },
...
         "SchemaUpdate": {
            "properties": {
               "name": {
                  "anyOf": [
                     {
                        "type": "string",
                        "maxLength": 50
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Name"
               },
               "version": {
                  "anyOf": [
                     {
                        "type": "string",
                        "maxLength": 50
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Version"
               },
               "data": {
                  "anyOf": [
                     {
                        "additionalProperties": true,
                        "type": "object"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Data"
               },
               "schema_entity_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Schema Entity Id"
               }
            },
            "type": "object",
            "title": "SchemaUpdate"
         },
         "Schemas": {
            "properties": {
               "name": {
                  "type": "string",
                  "maxLength": 50,
                  "title": "Name"
               },
               "version": {
                  "type": "string",
                  "maxLength": 50,
                  "title": "Version"
               },
               "data": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Data"
               },
               "schema_entity_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Schema Entity Id"
               },
               "id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Id"
               }
            },
            "type": "object",
            "required": [
               "name",
               "version"
            ],
            "title": "Schemas"
         },
...
         "ValidationError": {
            "properties": {
               "loc": {
                  "items": {
                     "anyOf": [
                        {
                           "type": "string"
                        },
                        {
                           "type": "integer"
                        }
                     ]
                  },
                  "type": "array",
                  "title": "Location"
               },
               "msg": {
                  "type": "string",
                  "title": "Message"
               },
               "type": {
                  "type": "string",
                  "title": "Error Type"
               },
               "input": {
                  "title": "Input"
               },
               "ctx": {
                  "type": "object",
                  "title": "Context"
               }
            },
            "type": "object",
            "required": [
               "loc",
               "msg",
               "type"
            ],
            "title": "ValidationError"
         }
      }
   }
}